bot/pyproject.toml
[project]
name = "claude-faq-bot"
version = "0.1.0"
description = "Discord bot wrapper for the Claude Community FAQ API"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.3",
    "httpx>=0.28.1",
]
